# openapi.yaml
openapi: 3.0.3
info:
  title: Flask TODO API
  description: |
    RESTful API for managing TODO items with JWT authentication.
  version: 1.0.0
  contact:
    name: API Support
    email: mail@barrydaniels.nl
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:5050/api/v1
    description: Development server
  - url: https://api.example.com/v1
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Authentication
    description: User registration and authentication
  - name: Todos
    description: TODO item management

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check if the API is running
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  service:
                    type: string
                    example: flask-todo-api

  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email, username, and password
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and return JWT tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using a valid refresh token
      operationId: refreshToken
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get the currently authenticated user's information
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /todos:
    get:
      tags:
        - Todos
      summary: List all todos
      description: Get a paginated list of todos for the current user
      operationId: listTodos
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: completed
          in: query
          description: Filter by completion status
          schema:
            type: boolean
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [title, completed, priority, due_date, created_at, updated_at]
            default: created_at
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Todos
      summary: Create a new todo
      description: Create a new todo item for the current user
      operationId: createTodo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /todos/{todoId}:
    get:
      tags:
        - Todos
      summary: Get a specific todo
      description: Get a todo by its ID
      operationId: getTodo
      security:
        - BearerAuth: []
      parameters:
        - name: todoId
          in: path
          required: true
          description: Todo ID
          schema:
            type: integer
      responses:
        '200':
          description: Todo found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Todos
      summary: Update a todo
      description: Update an existing todo by its ID
      operationId: updateTodo
      security:
        - BearerAuth: []
      parameters:
        - name: todoId
          in: path
          required: true
          description: Todo ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Todos
      summary: Delete a todo
      description: Delete a todo by its ID
      operationId: deleteTodo
      security:
        - BearerAuth: []
      parameters:
        - name: todoId
          in: path
          required: true
          description: Todo ID
          schema:
            type: integer
      responses:
        '200':
          description: Todo deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /todos/{todoId}/toggle:
    post:
      tags:
        - Todos
      summary: Toggle todo completion
      description: Toggle the completion status of a todo
      operationId: toggleTodo
      security:
        - BearerAuth: []
      parameters:
        - name: todoId
          in: path
          required: true
          description: Todo ID
          schema:
            type: integer
      responses:
        '200':
          description: Todo toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme

  schemas:
    UserRegister:
      type: object
      required:
        - email
        - username
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        username:
          type: string
          minLength: 3
          maxLength: 100
          description: Unique username
          example: johndoe
        password:
          type: string
          minLength: 8
          maxLength: 100
          description: User password (minimum 8 characters)
          example: securepassword123

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          description: User password
          example: securepassword123

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: User ID
          example: 1
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        username:
          type: string
          description: Username
          example: johndoe
        is_active:
          type: boolean
          description: Whether the user account is active
          example: true

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: JWT refresh token
        token_type:
          type: string
          description: Token type
          example: bearer

    TodoCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Todo title
          example: Buy groceries
        description:
          type: string
          maxLength: 1000
          description: Optional description
          example: Milk, eggs, bread
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
          description: Priority level
        due_date:
          type: string
          format: date-time
          description: Optional due date
          example: "2024-12-31T23:59:59Z"

    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Todo title
        description:
          type: string
          maxLength: 1000
          description: Optional description
        completed:
          type: boolean
          description: Completion status
        priority:
          type: string
          enum: [low, medium, high]
          description: Priority level
        due_date:
          type: string
          format: date-time
          description: Optional due date

    TodoResponse:
      type: object
      properties:
        id:
          type: integer
          description: Todo ID
          example: 1
        title:
          type: string
          description: Todo title
          example: Buy groceries
        description:
          type: string
          nullable: true
          description: Todo description
          example: Milk, eggs, bread
        completed:
          type: boolean
          description: Completion status
          example: false
        priority:
          type: string
          enum: [low, medium, high]
          description: Priority level
          example: medium
        due_date:
          type: string
          format: date-time
          nullable: true
          description: Due date
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        user_id:
          type: integer
          description: Owner user ID

    TodoListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TodoResponse'
          description: List of todos
        total:
          type: integer
          description: Total number of todos
          example: 42
        page:
          type: integer
          description: Current page number
          example: 1
        per_page:
          type: integer
          description: Items per page
          example: 10
        pages:
          type: integer
          description: Total number of pages
          example: 5

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Response message
          example: Operation completed successfully

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
          example: validation_error
        message:
          type: string
          description: Human-readable error message
          example: Validation failed
        details:
          type: object
          nullable: true
          description: Additional error details